Separates Login Page unit and integration tests.
