+++ test/integration/LoginPage.test.tsx
@@ -5,55 +5,11 @@
 import "@testing-library/jest-dom";
 
 import { render, screen, fireEvent } from "@testing-library/react";
-import LoginText from "../components/login/LoginText";
-import LoginTextBox from "../components/login/LoginTextBox";
-
-import Login from "../pages/Login";
+import Login from "../../src/pages/Login";
 import { BrowserRouter } from "react-router-dom";
 
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
-/*
-    Tests whether the LoginText component exists and renders correctly.
-    Checks whether the text appears correctly.
-*/
-describe("LoginText", () => {
-  it("renders the correct heading text", () => {
-    render(<LoginText />);
-
-    const heading = screen.getByRole("heading", {
-      name: /login to carelink!/i,
-    });
-
-    expect(heading).toBeInTheDocument();
-  });
-});
-
-/*
-
-    This tests the LoginTextBox component. It tests whether the component renders, the props were passed correctly,
-    and that the attributes were applied correctly.
-    
-*/
-
-describe("LoginTextBox", () => {
-  it("renders an input with correct attributes", () => {
-    render(
-      <LoginTextBox
-        name="username"
-        id="username"
-        placeholder="Enter username"
-      />,
-    );
-
-    const input = screen.getByPlaceholderText(/enter username/i);
-
-    expect(input).toBeInTheDocument();
-    expect(input).toHaveAttribute("name", "username");
-    expect(input).toHaveAttribute("id", "username");
-    expect(input).toHaveAttribute("type", "text");
-  });
-});
 
 /*
     This tests the Login Page. It checks whethers the navigation on button click works correctly.
